requirements.txt

langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
faiss-cpu>=1.7.4
streamlit>=1.29.0
python-dotenv>=1.0.0
openai>=1.3.0
matplotlib==3.7.2
seaborn==0.13.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
evaluate==0.4.1
bert-score==0.3.13
nltk==3.8.1
tiktoken==0.5.2
plotly==5.18.0
tenacity==8.2.3
pymupdf==1.23.7
tqdm==4.66.1